refactor: qbind.GetRequest() -> shared.GetRequest()
